Changed the EventType to double and minor change in RequestServiceTest

+++ Project1/src/main/java/com/revature/beans/EventType.java
@@ -1,21 +1,21 @@
 package com.revature.beans;
 
 public enum EventType {
-	UNIVERSITY_COURSE(0.8f), 
-	SEMINAR(0.6f), 
-	CERTIFICATION_PREPERATION_CLASS(0.75f),
-	CERTIFICATION(1f), 
-	TECHNICAL_TRAINING(0.9f), 
-	OTHER(0.3f);
+	UNIVERSITY_COURSE(0.8), 
+	SEMINAR(0.6), 
+	CERTIFICATION_PREPERATION_CLASS(0.75),
+	CERTIFICATION(1.0), 
+	TECHNICAL_TRAINING(0.9), 
+	OTHER(0.3);
 	
 	//How much of the total fee this course can be reimbursed for
-	private Float percent;
+	private Double percent;
 	
-	EventType(float percent){
+	EventType(Double percent){
 		this.percent = percent;
 	}
 	
-	public Float getPercent() {
+	public Double getPercent() {
 		return percent;
 	}
 }
